bigquery: add Client.list_rows, remove Table.fetch_data (#4119)
