+++ bigquery/tests/unit/test_table.py
@@ -945,286 +945,6 @@ class TestTable(unittest.TestCase, _SchemaBase):
         self.assertEqual(req['data'], SENT)
         self._verifyResourceProperties(table, RESOURCE)
 
-    def test_fetch_data_wo_schema(self):
-        from google.cloud.bigquery.table import _TABLE_HAS_NO_SCHEMA
-
-        client = _Client(project=self.PROJECT)
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        table = self._make_one(table_ref, client=client)
-
-        with self.assertRaises(ValueError) as exc:
-            table.fetch_data()
-
-        self.assertEqual(exc.exception.args, (_TABLE_HAS_NO_SCHEMA,))
-
-    def test_fetch_data_w_bound_client(self):
-        import datetime
-        import six
-        from google.cloud._helpers import UTC
-        from google.cloud.bigquery.table import SchemaField
-
-        PATH = 'projects/%s/datasets/%s/tables/%s/data' % (
-            self.PROJECT, self.DS_ID, self.TABLE_NAME)
-        WHEN_TS = 1437767599.006
-        WHEN = datetime.datetime.utcfromtimestamp(WHEN_TS).replace(
-            tzinfo=UTC)
-        WHEN_1 = WHEN + datetime.timedelta(seconds=1)
-        WHEN_2 = WHEN + datetime.timedelta(seconds=2)
-        ROWS = 1234
-        TOKEN = 'TOKEN'
-
-        def _bigquery_timestamp_float_repr(ts_float):
-            # Preserve microsecond precision for E+09 timestamps
-            return '%0.15E' % (ts_float,)
-
-        DATA = {
-            'totalRows': str(ROWS),
-            'pageToken': TOKEN,
-            'rows': [
-                {'f': [
-                    {'v': 'Phred Phlyntstone'},
-                    {'v': '32'},
-                    {'v': _bigquery_timestamp_float_repr(WHEN_TS)},
-                ]},
-                {'f': [
-                    {'v': 'Bharney Rhubble'},
-                    {'v': '33'},
-                    {'v': _bigquery_timestamp_float_repr(WHEN_TS + 1)},
-                ]},
-                {'f': [
-                    {'v': 'Wylma Phlyntstone'},
-                    {'v': '29'},
-                    {'v': _bigquery_timestamp_float_repr(WHEN_TS + 2)},
-                ]},
-                {'f': [
-                    {'v': 'Bhettye Rhubble'},
-                    {'v': None},
-                    {'v': None},
-                ]},
-            ]
-        }
-
-        conn = _Connection(DATA)
-        client = _Client(project=self.PROJECT, connection=conn)
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        full_name = SchemaField('full_name', 'STRING', mode='REQUIRED')
-        age = SchemaField('age', 'INTEGER', mode='NULLABLE')
-        joined = SchemaField('joined', 'TIMESTAMP', mode='NULLABLE')
-        table = self._make_one(table_ref, schema=[full_name, age, joined],
-                               client=client)
-
-        iterator = table.fetch_data()
-        page = six.next(iterator.pages)
-        rows = list(page)
-        total_rows = iterator.total_rows
-        page_token = iterator.next_page_token
-
-        self.assertEqual(len(rows), 4)
-        self.assertEqual(rows[0], ('Phred Phlyntstone', 32, WHEN))
-        self.assertEqual(rows[1], ('Bharney Rhubble', 33, WHEN_1))
-        self.assertEqual(rows[2], ('Wylma Phlyntstone', 29, WHEN_2))
-        self.assertEqual(rows[3], ('Bhettye Rhubble', None, None))
-        self.assertEqual(total_rows, ROWS)
-        self.assertEqual(page_token, TOKEN)
-
-        self.assertEqual(len(conn._requested), 1)
-        req = conn._requested[0]
-        self.assertEqual(req['method'], 'GET')
-        self.assertEqual(req['path'], '/%s' % PATH)
-
-    def test_fetch_data_w_alternate_client(self):
-        import six
-        from google.cloud.bigquery.table import SchemaField
-
-        PATH = 'projects/%s/datasets/%s/tables/%s/data' % (
-            self.PROJECT, self.DS_ID, self.TABLE_NAME)
-        MAX = 10
-        TOKEN = 'TOKEN'
-        DATA = {
-            'rows': [
-                {'f': [
-                    {'v': 'Phred Phlyntstone'},
-                    {'v': '32'},
-                    {'v': 'true'},
-                    {'v': '3.1415926'},
-                ]},
-                {'f': [
-                    {'v': 'Bharney Rhubble'},
-                    {'v': '33'},
-                    {'v': 'false'},
-                    {'v': '1.414'},
-                ]},
-                {'f': [
-                    {'v': 'Wylma Phlyntstone'},
-                    {'v': '29'},
-                    {'v': 'true'},
-                    {'v': '2.71828'},
-                ]},
-                {'f': [
-                    {'v': 'Bhettye Rhubble'},
-                    {'v': '27'},
-                    {'v': None},
-                    {'v': None},
-                ]},
-            ]
-        }
-        conn1 = _Connection()
-        client1 = _Client(project=self.PROJECT, connection=conn1)
-        conn2 = _Connection(DATA)
-        client2 = _Client(project=self.PROJECT, connection=conn2)
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        full_name = SchemaField('full_name', 'STRING', mode='REQUIRED')
-        age = SchemaField('age', 'INTEGER', mode='REQUIRED')
-        voter = SchemaField('voter', 'BOOLEAN', mode='NULLABLE')
-        score = SchemaField('score', 'FLOAT', mode='NULLABLE')
-        table = self._make_one(table_ref,
-                               schema=[full_name, age, voter, score],
-                               client=client1)
-
-        iterator = table.fetch_data(
-            client=client2, max_results=MAX, page_token=TOKEN)
-        page = six.next(iterator.pages)
-        rows = list(page)
-        total_rows = iterator.total_rows
-        page_token = iterator.next_page_token
-
-        self.assertEqual(len(rows), 4)
-        self.assertEqual(rows[0], ('Phred Phlyntstone', 32, True, 3.1415926))
-        self.assertEqual(rows[1], ('Bharney Rhubble', 33, False, 1.414))
-        self.assertEqual(rows[2], ('Wylma Phlyntstone', 29, True, 2.71828))
-        self.assertEqual(rows[3], ('Bhettye Rhubble', 27, None, None))
-        self.assertIsNone(total_rows)
-        self.assertIsNone(page_token)
-
-        self.assertEqual(len(conn1._requested), 0)
-        self.assertEqual(len(conn2._requested), 1)
-        req = conn2._requested[0]
-        self.assertEqual(req['method'], 'GET')
-        self.assertEqual(req['path'], '/%s' % PATH)
-        self.assertEqual(req['query_params'],
-                         {'maxResults': MAX, 'pageToken': TOKEN})
-
-    def test_fetch_data_w_repeated_fields(self):
-        import six
-        from google.cloud.bigquery.table import SchemaField
-
-        PATH = 'projects/%s/datasets/%s/tables/%s/data' % (
-            self.PROJECT, self.DS_ID, self.TABLE_NAME)
-        ROWS = 1234
-        TOKEN = 'TOKEN'
-        DATA = {
-            'totalRows': ROWS,
-            'pageToken': TOKEN,
-            'rows': [
-                {'f': [
-                    {'v': [{'v': 'red'}, {'v': 'green'}]},
-                    {'v': [{
-                        'v': {
-                            'f': [
-                                {'v': [{'v': '1'}, {'v': '2'}]},
-                                {'v': [{'v': '3.1415'}, {'v': '1.414'}]},
-                            ]}
-                    }]},
-                ]},
-            ]
-        }
-        conn = _Connection(DATA)
-        client = _Client(project=self.PROJECT, connection=conn)
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        color = SchemaField('color', 'STRING', mode='REPEATED')
-        index = SchemaField('index', 'INTEGER', 'REPEATED')
-        score = SchemaField('score', 'FLOAT', 'REPEATED')
-        struct = SchemaField('struct', 'RECORD', mode='REPEATED',
-                             fields=[index, score])
-        table = self._make_one(table_ref, schema=[color, struct],
-                               client=client)
-
-        iterator = table.fetch_data()
-        page = six.next(iterator.pages)
-        rows = list(page)
-        total_rows = iterator.total_rows
-        page_token = iterator.next_page_token
-
-        self.assertEqual(len(rows), 1)
-        self.assertEqual(rows[0][0], ['red', 'green'])
-        self.assertEqual(rows[0][1], [{'index': [1, 2],
-                                       'score': [3.1415, 1.414]}])
-        self.assertEqual(total_rows, ROWS)
-        self.assertEqual(page_token, TOKEN)
-
-        self.assertEqual(len(conn._requested), 1)
-        req = conn._requested[0]
-        self.assertEqual(req['method'], 'GET')
-        self.assertEqual(req['path'], '/%s' % PATH)
-
-    def test_fetch_data_w_record_schema(self):
-        import six
-        from google.cloud.bigquery.table import SchemaField
-
-        PATH = 'projects/%s/datasets/%s/tables/%s/data' % (
-            self.PROJECT, self.DS_ID, self.TABLE_NAME)
-        ROWS = 1234
-        TOKEN = 'TOKEN'
-        DATA = {
-            'totalRows': ROWS,
-            'pageToken': TOKEN,
-            'rows': [
-                {'f': [
-                    {'v': 'Phred Phlyntstone'},
-                    {'v': {'f': [{'v': '800'}, {'v': '555-1212'}, {'v': 1}]}},
-                ]},
-                {'f': [
-                    {'v': 'Bharney Rhubble'},
-                    {'v': {'f': [{'v': '877'}, {'v': '768-5309'}, {'v': 2}]}},
-                ]},
-                {'f': [
-                    {'v': 'Wylma Phlyntstone'},
-                    {'v': None},
-                ]},
-            ]
-        }
-        conn = _Connection(DATA)
-        client = _Client(project=self.PROJECT, connection=conn)
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        full_name = SchemaField('full_name', 'STRING', mode='REQUIRED')
-        area_code = SchemaField('area_code', 'STRING', 'REQUIRED')
-        local_number = SchemaField('local_number', 'STRING', 'REQUIRED')
-        rank = SchemaField('rank', 'INTEGER', 'REQUIRED')
-        phone = SchemaField('phone', 'RECORD', mode='NULLABLE',
-                            fields=[area_code, local_number, rank])
-        table = self._make_one(table_ref, schema=[full_name, phone],
-                               client=client)
-
-        iterator = table.fetch_data()
-        page = six.next(iterator.pages)
-        rows = list(page)
-        total_rows = iterator.total_rows
-        page_token = iterator.next_page_token
-
-        self.assertEqual(len(rows), 3)
-        self.assertEqual(rows[0][0], 'Phred Phlyntstone')
-        self.assertEqual(rows[0][1], {'area_code': '800',
-                                      'local_number': '555-1212',
-                                      'rank': 1})
-        self.assertEqual(rows[1][0], 'Bharney Rhubble')
-        self.assertEqual(rows[1][1], {'area_code': '877',
-                                      'local_number': '768-5309',
-                                      'rank': 2})
-        self.assertEqual(rows[2][0], 'Wylma Phlyntstone')
-        self.assertIsNone(rows[2][1])
-        self.assertEqual(total_rows, ROWS)
-        self.assertEqual(page_token, TOKEN)
-
-        self.assertEqual(len(conn._requested), 1)
-        req = conn._requested[0]
-        self.assertEqual(req['method'], 'GET')
-        self.assertEqual(req['path'], '/%s' % PATH)
-
     def test_row_from_mapping_wo_schema(self):
         from google.cloud.bigquery.table import _TABLE_HAS_NO_SCHEMA
         MAPPING = {'full_name': 'Phred Phlyntstone', 'age': 32}
